third_party/WebKit/LayoutTests/fast/css/invalidation/shadow-boundary-crossing-expected.txt

CONSOLE WARNING: line 1: /deep/ combinator is deprecated and will be a no-op in M60, around August 2017. See https://www.chromestatus.com/features/4964279606312960 for more details.
Invalidation sets should not apply across shadow roots if tree boundary crossing rules are not used.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.updateStyleAndReturnAffectedElementCount() is 10
PASS internals.updateStyleAndReturnAffectedElementCount() is 30
PASS successfullyParsed is true

TEST COMPLETE
